feat(docs): sync API reference into DigiVault
Generate Obsidian-style vault notes for the API docs so the digithings.ai chat
can search and cite them. The authored /docs content stays the single source of
truth: scripts/gen-api-vault.ts reuses the shared markdown serializers (extracted
to lib/docsSerializers.ts) to emit one note per module and per guide into
docs/vision/api/. The existing production sync (sync_architecture_vault.py, on
push to main) upserts them into Supabase architecture_notes — no new write path.

- New: scripts/gen-api-vault.ts, lib/docsSerializers.ts, docs/vision/api/*.md (13).
- DocsLayout imports the serializers (no behavior change; deduped logic).
- make gen-api-vault regenerates; validated via the sync --dry-run.

Fixes #1167
